+++ scripts/repo/winesapos-build-repo.sh
@@ -82,6 +82,7 @@ cd "${WORK_DIR}"
 makepkg_fn apfsprogs-git
 makepkg_fn ayaneo-platform-dkms-git
 makepkg_fn ayaled-updated
+makepkg_fn bmi260-dkms
 makepkg_fn ceph-bin
 makepkg_fn curl-static-bin
 makepkg_fn gfs2-utils